requirements.txt
Django==2.2.4
ConcurrentLogHandler
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pyotp==2.3.0
urllib3==1.25.7
uWSGI==2.0.18
requests==2.22.0
psycopg2-binary
django-filter==2.3.0
django-cors-headers==3.1.0
